requirements.txt
pandas==1.3.0
numpy==1.18.5
imbalanced_learn==0.8.0
scipy==1.7.0
lightgbm==3.3.0
imblearn==0.0
scikit_learn==1.0.2
lime==0.2.0.1
shap==0.40.0
ipykernel
PyYAML
dill
